Add sound. Create sound wav at http://www.superflashbros.net/as3sfxr/
Fix position of options cell knowing that check box contains 2 cells - image and label. So we pad them accordingly so that the check is nicely position on the left.

Remove unused images and keep used ui skin only. Skin use produce less images and we can add effects on click or hover.